// web/src/api/generated/models/UploadStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum UploadStatus {
    CREATED = 'CREATED',
    UPLOADING = 'UPLOADING',
    COMPLETING = 'COMPLETING',
    COMPLETED = 'COMPLETED',
    FAILED = 'FAILED',
    EXPIRED = 'EXPIRED',
}